pyproject.toml
[tool.poetry]
name = "fastapi-realworld-example-app"
version = "0.1.0"
description = ""
authors = ["Adrien Beaudouin <adrien.beaudouin@outlook.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = { extras = ["standard"], version = "^0.23.0" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
alembic = "^1.8.1"
python-slugify = "^8.0.0"
pydantic = { extras = ["email"], version = "^2.0.0" }
requests = "^2.28.1"
psycopg2-binary = "^2.9.5"
httpx = "^0.25.0"
sqlalchemy = { extras = ["mypy"], version = "^2.0.7" }
pydantic-settings = "^2.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
pytest = "^7.1.2"
pytest-cov = "^4.0.0"
faker = "^19.0.0"
types-python-slugify = "^8.0.0"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120
extend-select = ["UP", "RUF", "I"]

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"
ignore_missing_imports = true
disallow_untyped_defs = true
